+++ lib/src/weights.dart
@@ -1,3 +1,5 @@
+import 'dart:math' as math;
+
 import 'package:smollm2/src/quant_type.dart';
 
 import 'config.dart';
@@ -27,23 +29,40 @@ class ModelWeights {
   late List<LayerWeights> layers;
   late FP32Tensor finalNorm;
 
-  void load(Config config, DataReader dataReader) {
-    _loadEmbedTokens(config, dataReader);
-    _loadLayers(config, dataReader);
+  void load(
+    Config config,
+    DataReader dataReader, {
+    math.Random? jitterRandom,
+    double? jitterScale,
+  }) {
+    _loadEmbedTokens(config, dataReader, jitterRandom, jitterScale);
+    _loadLayers(config, dataReader, jitterRandom, jitterScale);
     _loadFinalNorm(config, dataReader);
   }
 
-  void _loadEmbedTokens(Config config, DataReader dataReader) {
-    embedTokens = _readQ(
+  void _loadEmbedTokens(
+    Config config,
+    DataReader dataReader,
+    math.Random? jitterRandom,
+    double? jitterScale,
+  ) {
+    embedTokens = _readTensor(
       dataReader,
       config.vocabSize,
       config.hiddenSize,
+      jitterRandom: jitterRandom,
+      jitterScale: jitterScale,
       quantType: config.quantType,
       fp32: true,
     );
   }
 
-  void _loadLayers(Config config, DataReader dataReader) {
+  void _loadLayers(
+    Config config,
+    DataReader dataReader,
+    math.Random? jitterRandom,
+    double? jitterScale,
+  ) {
     final quantType = config.quantType;
     final hd = config.headDim;
 
@@ -56,34 +75,42 @@ class ModelWeights {
 
       ////
 
-      lw.qProj = _readQ(
+      lw.qProj = _readTensor(
         dataReader,
         config.numHeads * hd,
         config.hiddenSize,
+        jitterRandom: jitterRandom,
+        jitterScale: jitterScale,
         quantType: quantType,
         fp32: true,
       );
 
-      lw.kProj = _readQ(
+      lw.kProj = _readTensor(
         dataReader,
         config.numKvHeads * hd,
         config.hiddenSize,
+        jitterRandom: jitterRandom,
+        jitterScale: jitterScale,
         quantType: quantType,
         fp32: true,
       );
 
-      lw.vProj = _readQ(
+      lw.vProj = _readTensor(
         dataReader,
         config.numKvHeads * hd,
         config.hiddenSize,
+        jitterRandom: jitterRandom,
+        jitterScale: jitterScale,
         quantType: quantType,
         fp32: true,
       );
 
-      lw.oProj = _readQ(
+      lw.oProj = _readTensor(
         dataReader,
         config.hiddenSize,
         config.numHeads * hd,
+        jitterRandom: jitterRandom,
+        jitterScale: jitterScale,
         quantType: quantType,
         fp32: true,
       );
@@ -97,26 +124,32 @@ class ModelWeights {
 
       ////
 
-      lw.gateProj = _readQ(
+      lw.gateProj = _readTensor(
         dataReader,
         config.intermediateSize,
         config.hiddenSize,
+        jitterRandom: jitterRandom,
+        jitterScale: jitterScale,
         quantType: quantType,
         fp32: true,
       );
 
-      lw.upProj = _readQ(
+      lw.upProj = _readTensor(
         dataReader,
         config.intermediateSize,
         config.hiddenSize,
+        jitterRandom: jitterRandom,
+        jitterScale: jitterScale,
         quantType: quantType,
         fp32: true,
       );
 
-      lw.downProj = _readQ(
+      lw.downProj = _readTensor(
         dataReader,
         config.hiddenSize,
         config.intermediateSize,
+        jitterRandom: jitterRandom,
+        jitterScale: jitterScale,
         quantType: quantType,
         fp32: true,
       );
@@ -129,38 +162,78 @@ class ModelWeights {
     finalNorm = FP32Tensor.readFrom(dataReader, config.hiddenSize);
   }
 
-  Tensor _readQ(
+  Tensor _readTensor(
     DataReader br,
     int rows,
     int cols, {
     required QuantType quantType,
     bool fp32 = false,
+    math.Random? jitterRandom,
+    double? jitterScale,
   }) {
     switch (quantType) {
       case QuantType.q8:
-        return _readQ8(br, rows, cols, fp32: fp32);
+        return _readQ8(
+          br,
+          rows,
+          cols,
+          jitterRandom: jitterRandom,
+          jitterScale:
+              jitterScale ?? Q8Tensor.defaultQ8DequantizationJitterScale,
+          fp32: fp32,
+        );
       case QuantType.q16:
-        return _readQ16(br, rows, cols, fp32: fp32);
+        return _readQ16(
+          br,
+          rows,
+          cols,
+          jitterRandom: jitterRandom,
+          jitterScale:
+              jitterScale ?? Q16Tensor.defaultQ16DequantizationJitterScale,
+          fp32: fp32,
+        );
       default:
         throw UnsupportedError("quantType: $quantType");
     }
   }
 
-  Tensor _readQ8(DataReader br, int rows, int cols, {bool fp32 = false}) {
+  Tensor _readQ8(
+    DataReader br,
+    int rows,
+    int cols, {
+    bool fp32 = false,
+    math.Random? jitterRandom,
+    double jitterScale = Q8Tensor.defaultQ8DequantizationJitterScale,
+  }) {
     final q8 = Q8Tensor.readFrom(br, rows, cols);
 
     if (fp32) {
-      return q8.toFP32Tensor(cached: false);
+      return q8.toFP32Tensor(
+        cached: false,
+        jitterRandom: jitterRandom,
+        jitterScale: jitterScale,
+      );
     }
 
     return q8;
   }
 
-  Tensor _readQ16(DataReader br, int rows, int cols, {bool fp32 = false}) {
+  Tensor _readQ16(
+    DataReader br,
+    int rows,
+    int cols, {
+    bool fp32 = false,
+    math.Random? jitterRandom,
+    double jitterScale = Q16Tensor.defaultQ16DequantizationJitterScale,
+  }) {
     final q16 = Q16Tensor.readFrom(br, rows, cols);
 
     if (fp32) {
-      return q16.toFP32Tensor(cached: false);
+      return q16.toFP32Tensor(
+        cached: false,
+        jitterRandom: jitterRandom,
+        jitterScale: jitterScale,
+      );
     }
 
     return q16;